requirements.txt
quart>=0.18.4
numpy>=1.24.0
pandas>=2.0.0
openai>=1.0.0
pinecone>=2.2.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
tqdm>=4.65.0
google-generativeai>=0.3.0
scikit-learn>=1.3.0 